pyproject.toml
[tool.poetry]
name = "datajobs"
version = "0.1.0"
description = ""
authors = ["Giancarlo Lester <gian.lepear@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.2.3"
djangorestframework = "^3.12.4"
beautifulsoup4 = "^4.9.3"
pandas = "^1.2.4"
python-decouple = "^3.4"
psycopg2-binary = "^2.8.6"

[tool.poetry.dev-dependencies]
jupyterlab = "^3.4.4"
black = "^22.6.0"
isort = "^5.10.1"
flake8 = "^4.0.1"


[tool.black]
line-length = 90
target-version = ['py38']
include = '\.pyi?$'
exclude = '''

(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
    | migrations
    | alembic
    | postgres-data
  )/
  | foo.py           # also separately exclude a file named foo.py in
                     # the root of the project
)
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 90
skip_gitignore = "true"
skip_glob = "**/alembic/**"
skip =  "settings.py"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

